Inject thread detail data source
