+++ iOS/Charter/Charter/Thread Detail View Controller/ThreadDetailViewController.swift
@@ -8,135 +8,18 @@
 
 import UIKit
 
-protocol TableViewCellIndentationLevelDataSource: class {
-    func tableView(tableView: UITableView, indentationLevelForRowAtIndexPath indexPath: NSIndexPath) -> Int
-}
-
-class ThreadDetailDataSource: NSObject, UITableViewDataSource, TableViewCellIndentationLevelDataSource {
-    private let service: EmailThreadService
-    weak var cellDelegate: FullEmailMessageTableViewCellDelegate?
-    
-    private let cellIdentifier = "emailCell"
-    
-    private var indentationAndEmail: [(Int, Email)] = [] {
-        didSet {
-            textViewDataSources = [NSIndexPath: EmailCollapsibleTextViewDataSource]()
-        }
-    }
-    
-    private let rootEmail: Email
-    private var textViewDataSources: [NSIndexPath: EmailCollapsibleTextViewDataSource] = [NSIndexPath: EmailCollapsibleTextViewDataSource]()
-    private lazy var emailFormatter: EmailFormatter = EmailFormatter()
-    
-    private var emails: [Email] = [] {
-        didSet {
-            self.indentationAndEmail = self.computeIndentationLevels(rootEmail)
-        }
-    }
-    
-    init(tableView: UITableView, service: EmailThreadService, rootEmail: Email) {
-        self.service = service
-        self.rootEmail = rootEmail
-        tableView.registerNib(FullEmailMessageTableViewCell.nib(), forCellReuseIdentifier: cellIdentifier)
-        super.init()
-        
-        service.getCachedThreads(descendantsRequestForRootEmail(rootEmail)) { (emails) -> Void in
-            self.emails = emails
-            tableView.reloadData()
-            
-            if rootEmail.descendants.count > emails.count {
-                // Get uncached threads if we are missing any
-                service.getUncachedThreads(self.descendantsRequestForRootEmail(rootEmail), completion: { (descendants) -> Void in
-                    self.emails = descendants
-                    tableView.reloadData()
-                })
-            }
-        }
-    }
-    
-    func tableView(tableView: UITableView, indentationLevelForRowAtIndexPath indexPath: NSIndexPath) -> Int {
-        return indentationAndEmail[indexPath.row].0
-    }
-    
-    func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCellWithIdentifier(cellIdentifier) as! FullEmailMessageTableViewCell
-        let email = indentationAndEmail[indexPath.row].1
-        
-        cell.indentationLevel = indentationAndEmail[indexPath.row].0
-        cell.indentationWidth = 10
-        cell.dateLabel.text = emailFormatter.formatDate(email.date)
-        cell.nameLabel.text = emailFormatter.formatName(email.from)
-        cell.delegate = cellDelegate
-        
-        var textViewDataSource = textViewDataSources[indexPath]
-        
-        let content = emailFormatter.formatContent(email.content)
-        
-        if textViewDataSource == nil {
-            let regions = EmailCollapsibleTextViewDataSource.QuoteRanges(content)
-            textViewDataSource = EmailCollapsibleTextViewDataSource(text: content, initiallyCollapsedRegions: regions)
-            textViewDataSources[indexPath] = textViewDataSource!
-        }
-        
-        cell.textViewDataSource = textViewDataSource!
-        return cell
-    }
-    
-    func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return indentationAndEmail.count
-    }
-    
-    private func descendantsRequestForRootEmail(rootEmail: Email) -> EmailThreadRequest {
-        let builder = EmailThreadRequestBuilder()
-        builder.idIn = Array(rootEmail.descendants.map { $0.id })
-        builder.page = 1
-        builder.pageSize = 1000
-        builder.onlyComplete = true
-        return builder.build()
-    }
-    
-    private func computeIndentationLevels(rootEmail: Email) -> [(Int, Email)] {
-        var children = [String: [Email]]()
-        for email in rootEmail.descendants {
-            if let inReplyTo = email.inReplyTo {
-                if children[inReplyTo.id] == nil {
-                    children[inReplyTo.id] = []
-                }
-                
-                children[inReplyTo.id]!.append(email)
-            }
-        }
-        
-        func indentationLevel(root: Email, indentLevel: Int) -> [(Int, Email)] {
-            var list = [(Int, Email)]()
-            for child in children[root.id] ?? [] {
-                if child.id != root.id {
-                    list.appendContentsOf(indentationLevel(child, indentLevel: indentLevel + 1))
-                }
-            }
-            return [(indentLevel, root)] + list
-        }
-        
-        let thread = indentationLevel(rootEmail, indentLevel: 0)
-        return thread
-    }
-}
-
 class ThreadDetailViewController: UIViewController, UITableViewDelegate, FullEmailMessageTableViewCellDelegate, UIPopoverPresentationControllerDelegate {
     @IBOutlet weak var tableView: UITableView!
     
-    private var dataSource: ThreadDetailDataSource!
-    private let rootEmail: Email
-    private let service: EmailThreadService
+    private let dataSource: ThreadDetailDataSource
     
-    init(service: EmailThreadService, rootEmail: Email) {
-        self.service = service
-        self.rootEmail = rootEmail
+    init(dataSource: ThreadDetailDataSource) {
+        self.dataSource = dataSource
         super.init(nibName: "ThreadDetailViewController", bundle: NSBundle.mainBundle())
     }
     
     override func viewDidLoad() {
-        dataSource = ThreadDetailDataSource(tableView: tableView, service: service, rootEmail: rootEmail)
+        dataSource.registerTableView(tableView)
         dataSource.cellDelegate = self
         tableView.dataSource = dataSource
         tableView.delegate = self
@@ -151,7 +34,7 @@ class ThreadDetailViewController: UIViewController, UITableViewDelegate, FullEma
     }
     
     func tableView(tableView: UITableView, indentationLevelForRowAtIndexPath indexPath: NSIndexPath) -> Int {
-        return dataSource?.tableView(tableView, indentationLevelForRowAtIndexPath: indexPath) ?? 0
+        return dataSource.tableView(tableView, indentationLevelForRowAtIndexPath: indexPath) ?? 0
     }
     
     func didChangeCellHeight(indexPath: NSIndexPath) {
